fixed rendering of a's, but not b or d or f (and most glyphs)

+++ font-rendeirng/font-rendering/renderFace.hpp
@@ -20,6 +20,6 @@ enum class Segment {
 };
 
 simd_float2 getMidpoint(simd_float2 pointA, simd_float2 pointB);
-std::vector<simd_float2> drawBezier(Segment segment, std::vector<simd_float2> buffer, int resolution);
-std::vector<simd_float2> drawContour(FT_Vector* points, unsigned char* tags, long start, long end, int resolution);
-std::vector<std::vector<simd_float2>> drawContours(char ch, FT_Library ft, std::string_view fontPath, int resolution);
+std::vector<simd_float2> drawBezier(Segment segment, std::vector<simd_float2> buffer, float resolution);
+std::vector<simd_float2> drawContour(FT_Vector* points, unsigned char* tags, long start, long end, float resolution);
+std::vector<std::vector<simd_float2>> drawContours(char ch, FT_Library ft, std::string_view fontPath, float resolution);
